+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Updated eslint to 9.35.0
+- Enabled preserve-caught-error rule with `{ requireCatchParameter: false }`
+
 
 ## [45.0.0] - 2025-09-04